Impl observer registration in LocalAudioSource
